requirements.txt
sievedata==0.0.27
requests==2.31.0 